add time checking

+++ src/shift/shift.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from "@nestjs/common"
+import {
+    BadRequestException,
+    Injectable,
+    NotFoundException,
+} from "@nestjs/common"
 import { Repository, Between } from "typeorm"
 import { InjectRepository } from "@nestjs/typeorm"
 import { Shift } from "./shift.entity"
@@ -43,20 +47,73 @@ export class ShiftService {
         return data
     }
 
-    create(body: CreateShiftDto) {
+    async findByDatetime(body: Partial<CreateShiftDto>) {
+        const date = moment(body.date, "yyyy-MM-DD")
+        const start_date = date.clone().subtract(1, "day")
+        const end_date = date.clone().add(1, "day")
+
+        const data = await this.repo.find({
+            where: {
+                date: Between(start_date.toDate(), end_date.toDate()),
+            },
+        })
+        let valid = true
+        if (data) {
+            data.forEach((el) => {
+                const start_time = moment(el.start_time, "HH:mm").subtract(
+                    "2",
+                    "minutes",
+                )
+                const end_time = moment(el.end_time, "HH:mm").add(
+                    "2",
+                    "minutes",
+                )
+                const m_start = moment(body.start_time, "HH:mm")
+                const m_end = moment(body.end_time, "HH:mm")
+                if (
+                    m_start.isBetween(start_time, end_time) ||
+                    m_end.isBetween(start_time, end_time)
+                ) {
+                    valid = false
+                }
+            })
+        }
+        return valid
+    }
+
+    async create(body: CreateShiftDto) {
+        const check = await this.findByDatetime(body)        
+        if (!check) {
+            throw new BadRequestException(
+                "Create Error, the date and time are clashing each other",
+            )
+        }
         const data = this.repo.create(body)
         return this.repo.save(data)
     }
 
     async update(id: number, body: Partial<CreateShiftDto>) {
+        if (body.start_time || body.end_time) {
+            const check = this.findByDatetime(body)
+            if (!check) {
+                throw new BadRequestException(
+                    "Create Error, the date and time are clashing each other",
+                )
+            }
+        }
         const data = await this.findOne(id)
-
+        if (data.is_published) {
+            throw new BadRequestException("update error, data has been publish")
+        }
         Object.assign(data, body)
         return this.repo.save(data)
     }
 
     async remove(id: number) {
         const data = await this.findOne(id)
+        if (data.is_published) {
+            throw new BadRequestException("delete error, data has been publish")
+        }
         return this.repo.remove(data)
     }
 }
